floats

+++ jsonprint.h
@@ -10,6 +10,6 @@
 
 FILE* openFile(char* fileName);
 void closeFile(FILE* file);
-void writeModelToFile(FILE* file, Timestamps timestamps, Selected_model model, int *first, int start_time, int end_time, double error, int column_id);
+void writeModelToFile(FILE* file, Timestamps timestamps, Selected_model model, int *first, int start_time, int end_time, float error, int column_id);
 void write_text_to_file(FILE* file, int* first, int column_id, int count, char* string, long start_time, long end_time, Timestamps timestamps);
 #endif
